% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxPoints}
\alias{maxPoints}
\title{The maximum number of points on a line}
\usage{
maxPoints(points)
}
\arguments{
\item{points}{Two dimensional matrix, the first column is x and the second column is y}
}
\value{
the mximum number
}
\description{
Find the maximum number of points on the two-dimensional plane that can be connected into a straight line
}
\examples{
\dontrun{
points<-matrix(c(1,1,2,2,3,3),3,2,byrow = TRUE)
maxPoints(points)
}
}
