% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{SGD_min}
\alias{SGD_min}
\title{use SGD to estimate the x0 where f(x0) get the min value.}
\usage{
SGD_min(x, alpha, N, coefficient, powers)
}
\arguments{
\item{x}{a point in the axis x}

\item{alpha}{each-iter moving step.}

\item{N}{Total number of iterations.}

\item{coefficient}{coefficient of function.}

\item{powers}{power of every variable in the function.}
}
\value{
the  x0 point .
}
\description{
estimating regression parameter by SGD.
}
\examples{
\dontrun{
x<-stats::rnorm(1,0,100);N<-2000
alpha<-0.005;coefficient<-c(1,3);powers<-c(2,4)
x_last<-SGD_min(x,alpha,N,coefficient,powers)
paste("The x corresponding to the function value is",x_last)
}
}
